# trimbledrive/docs/apispec.yaml
...
openapi: 3.0.1
info:
  title: Trimble Drive API
  description: |-
    The service provides capabilities for managing files in the context of Connected Data Environment
    serving the Industry Clouds needs.
    A "file" is an unstructured blob of information with set of standard and custom metadata attributes.

    **PRIOR TO USING THIS API, YOU MUST AGREE TO OUR [TERMS OF USE](https://connect.trimble.com/terms_of_service.html)**

    # Main Capabilities

    * Managing hierarchical file structures (files organized in to hierarchies with folders)
        * The "file" is an unstructured BLOB content with the metadata end user usually expects from the familiar
    "file" concept
        * File hierarchies are organized as "file spaces" to provide logical isolation. The file spaces can be visible
    to end users or completely hidden and used by internal systems to manage unstructured data/files
    * Basic file and folder operations: create/upload, move/rename, delete, copy
        * hierarchical/recursive operations (e.g. delete folder) are implemented asynchronously
        * concurrency control on all operations including upsert, insert-only and update-only semantics
    * Strict versioning for files and folders. All changes are trackable and addressable. User is able to see the
    state of the full folder structure as it was at any moment of the history. User is able to refer to any specific
    version of the file.
...
    (even those not visible to the user).
...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...
security:
  - Bearer: [ ]
...
paths:
...
  /spaces/{spaceId}/files/{fileId}:
    get:
      tags:
        - Files
      summary: Get File Details
      description: "This API will fetch the latest version details for the given file\
          \ Id."
      operationId: getFileDetails
      parameters:
        - name: spaceId
          in: path
          description: filespace ID
          required: true
          schema:
            type: string
        - name: fileId
          in: path
          required: true
          schema:
            type: string
        - name: deleted
          in: query
          description: "If specified and true, resource is returned even if it is marked as deleted. Otherwise 404 error returned for deleted resource."
          required: false
          schema:
            type: boolean
            default: false
        - name: fields
          in: query
          required: false
          explode: false
          description: Option to specify fields to be included in the response.
          schema:
            type: array
            items:
              type: string
              enum:
                - downloadUrl
                - name
                - size
                - parentFolderId
                - topActiveParentFolderId
        - name: urlExpiryInMins
          in: query
          description: Number of minutes for which the pre-signed download URL for the file will be valid.
          schema:
            type: integer
            default: 5
            minimum: 1
            maximum: 720
      responses:
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "200":
          description: OK
          headers:
            ETag:
              description: "Weak ETag that can be used for the concurrency control."
              schema:
                type: "string"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileDetailsResponse'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
        - "Files"
      summary: "Move/Rename/Undelete File"
      description: "You can rename the file using this API. To move the file, change\
          \ the `parentFolderId` to the id of the destination folder. \
          \ \n\n To undelete a deleted file perform this API call with \
          \ empty request body and deleted param as true.\n\nThe API\
          \ also supports **If-Match** optional header to provide the version, so\
          \ you can verify you are working on the latest version.\n\nIf version provided\
          \ in the **If-Match** header is not latest, error Response with Response\
          \ code **412** will be send back to the caller with **Errorcode** INVALID_HEADER\
          \ &\n**Error Message**: If-Match header value is not latest version.\n\n\
          If the **If-Match** header value is *  or header not provided, actions will\
          \ be performed on the latest version."
      operationId: updateFileDetails
      parameters:
        - name: spaceId
          in: path
          description: filespace ID
          required: true
          schema:
            type: string
        - name: "fileId"
          in: "path"
          description: "ID of the File"
          required: true
          schema:
            type: "string"
        - name: deleted
          in: query
          required: false
          description: To be passed with value as `true` for restoring a deleed file.
          schema:
            type: boolean
            default: false
        - name: fields
          in: query
          required: false
          explode: false
          description: Option to specify fields to be included in the response.
          schema:
            type: array
            items:
              type: string
              enum:
                - downloadUrl
                - name
                - size
                - parentFolderId
                - topActiveParentFolderId
        - name: urlExpiryInMins
          in: query
          description: Number of minutes for which the pre-signed download URL for the file will be valid.
          schema:
            type: integer
            default: 5
            minimum: 1
            maximum: 720
        - name: "If-Match"
          in: "header"
          schema:
            type: "string"
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFileDetailsRequest"
        required: true
      responses:
        "200":
          description: ""
          headers:
            ETag:
              description: "Weak ETag that can be used for the concurrency control."
              schema:
                type: "string"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileDetailsResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Files
      summary: "Delete File"
      description: "The API also supports **If-Match** optional header to provide\
          \ the version, so you can verify you are working on the latest version.\n\n\
          If version provided in the **If-Match** header is not latest, error Response\
          \ wiith Response code **412** will be send back to the caller with **Errorcode**\
          \ INVALID_HEADER &\n**Error Message**: If-Match header value is not latest\
          \ version.\n\nIf the **If-Match** header value is *  or header not provided,\
          \ actions will be performed on the latest version."
      operationId: deleteFile
      parameters:
        - name: spaceId
          in: path
          description: filespace ID
          required: true
          schema:
            type: string
        - name: "fileId"
          in: "path"
          description: "ID of the file"
          required: true
          schema:
            type: "string"
        - name: "If-Match"
          in: "header"
          schema:
            type: "string"
        - name: fields
          in: query
          required: false
          explode: false
          description: Option to specify fields to be included in the response.
          schema:
            type: array
            items:
              type: string
              enum:
                - downloadUrl
                - name
                - size
                - parentFolderId
                - topActiveParentFolderId
      responses:
        "200":
          description: ""
          headers:
            ETag:
              description: "Weak ETag that can be used for the concurrency control."
              schema:
                type: "string"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileDetailsResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Permission denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ErrorResponse:
      title: "The error response."
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: "The error code. A non-empty language-independent string."
        message:
          type: string
          description: A human-readable string describing the error. The `Content-Language` header MUST contain the language code from [RFC5646](https://docs.oasis-open.org/odata/odata-json-format/v4.01/odata-json-format-v4.01.html#rfc5646) corresponding to the language in which the value for message is written.
...
    UpdateFileDetailsRequest:
      title: "UpdateFileDetailsRequest"
      type: "object"
      properties:
        name:
          type: "string"
          minLength: 1
          maxLength: 255
          description: "The new name."
        parentFolderId:
          type: "string"
          description: "The new parent folder identifier."
    FileDetailsResponse:
      type: object
      required:
        - id
        - version
        - type
        - spaceId
        - createdBy
        - modifiedBy
        - createdOn
        - modifiedOn
      properties:
        id:
          type: string
        version:
          type: string
          pattern: '^([1-9]\d*)\.([1-9]\d*)$'
        parentFolderId:
          type: string
        name:
          type: string
        type:
          type: string
          enum:
            - FILE
        spaceId:
          type: string
        size:
          type: integer
          format: int64
        createdBy:
          type: string
        modifiedBy:
          type: string
        createdOn:
          type: string
          format: date-time
        modifiedOn:
          type: string
          format: date-time
        deleted:
          type: boolean
          description: Deleted state flag
        downloadUrl:
          type: string
          format: url
          description: "Pre-signed URL for downloading the file content. \
                \ Part downloads supported with range headers. Use GET on this url \
                \ to download file content (no other parameters/headers required). \
                \ The headers: content-type, content-disposition and cache-control will be part \
                \ of the response while making the GET call on the pre-signed download url."
        topActiveParentFolderId:
          type: string
          description: Id of the topmost parent folder which is not in deleted state. Normally this will be equal to space `rootFolderId` for not deleted resources and equal to `null` (missing in the response) for deleted resources. If this property is not `null` and not equal to space `rootFolderId` then this resource belongs to orphan subtree with not deleted resources that is not connected to the space root folder.
...
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT